refactor ModelPredictiveController to by example agnostic
